refactor(kornia-py): extract DLPack capsule machinery into cuda_ext/capsule.rs
Second step of breaking up the oversized cuda.rs: move the DLPack capsule
pack/export (arc_dlpack_capsule*, wrap/build helpers, the destructor macro) and
the device import (dlpack_to_device_arc, PyKeepalive) into cuda_ext/capsule.rs,
re-exported from mod.rs so crate::cuda_ext::dlpack_to_device_arc and the
Tensor::__dlpack__ packers resolve unchanged. Named `capsule` to avoid clashing
with the crate-level `crate::dlpack` module.

cuda.rs (1826 lines) is now mod.rs (1270) + capsule.rs (354) + cuda_color.rs
(255), each a focused concern. No behavior change; verified on both feature
configs (build + clippy) and the device/DLPack/interop suites (63 passed, only
the known Jetson mem-probe noise-fail).
